System.Log.Handler.Growl

Description

Simple log handlers

Written by Richard M. Neswold, Jr. rich.neswold@gmail.com

Documentation

addTarget :: HostName -> GrowlHandler -> IO GrowlHandler

Adds a remote machine's address to the list of targets that will receive log messages. Calling this function sends a registration packet to the machine. This function will throw an exception if the host name cannot be found.

growlHandler

Arguments

:: String
The name of the service

-> Priority
Priority of handler

-> IO GrowlHandler

Creates a Growl handler. Once a Growl handler has been created, machines that are to receive the message have to be specified.